+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.13] - 2026-04-23
+
+### Docs
+- Update README.md
+- Update godot/README.md
+
+### Test
+- Update tests/smoke.sh
+
+### Other
+- Update godot/Dockerfile
+- Update godot/Makefile
+- Update godot/docker-compose.yml
+- Update godot/scripts/quickstart.sh
+- Update godot/scripts/test-services.sh
+
 ## [0.1.12] - 2026-04-23
 
 ### Docs